+++ tests/kernel/boo/conv_exports/cache_prepopulator_test.py
@@ -1,9 +1,13 @@
-import os
+# Copyright 2025 The IREE Authors
+#
+# Licensed under the Apache License v2.0 with LLVM Exceptions.
+# See https://llvm.org/LICENSE.txt for license information.
+# SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+
 from pathlib import Path
 from tempfile import TemporaryDirectory
 import unittest
 from iree.turbine.kernel.boo.runtime import set_cache_dir, clear_cache
-from iree.turbine.support.logging import runtime_logger as logger
 
 
 class CachePopulatorTest(unittest.TestCase):
@@ -12,14 +16,18 @@ class CachePopulatorTest(unittest.TestCase):
             cache_dir = Path(td)
             set_cache_dir(cache_dir=cache_dir)
 
-            from iree.turbine.kernel.boo.conv_exports import CachePopulator
+            from iree.turbine.kernel.boo.driver.preload import CachePopulator
+            from iree.turbine.kernel.boo.conv_exports.miopen_parser import ConvParser
+            from iree.turbine.kernel.boo.conv_exports.conv import ConvSignature
 
             commands = [
                 "convbfp16 -n 128 -c 128 -H 24 -W 48 -k 384 -y 1 -x 1 -p 0 -q 0 -u 1 -v 1 -l 1 -j 1 -m conv -g 1 -F 1 -t 1 --iter 100 --in_layout NHWC --out_layout NHWC --fil_layout NHWC",
                 "convbfp16 -n 128 -c 35 -H 48 -W 32 -k 35 -y 1 -x 1 -p 0 -q 0 -u 1 -v 1 -l 1 -j 1 -m conv -g 1 -F 1 -t 1 --iter 100 --in_layout NHWC --out_layout NHWC --fil_layout NHWC",
             ]
 
-            pop = CachePopulator(commands=commands)
+            pop = CachePopulator(
+                commands=commands, parser_cls=ConvParser, signature_cls=ConvSignature
+            )
 
             pop.run()
